requirements.txt.txt
streamlit
torch
numpy
pandas
matplotlib
Pillow
scikit-learn
